commonize handlers
